More wall smoothing "improvements"
Still some bugs to be worked out related to the turn code, but it works
most of the time(tm).

Updated the wall smoothing code to use a different implementation by
Voidious, and added a Javadoc header so that I actually know what to
pass in for the angle (which I spent around three hours trying to figure
out).